# approval-api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PeopleCore 전자결재 API",
    "description": "전자결재 시스템 (Approval) API 명세서 - 53개 엔드포인트",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development"
    }
  ],
...
  "paths": {
...
    "/approval/auto-classify-rules": {
      "get": {
        "tags": ["자동분류규칙"],
        "summary": "자동 분류 규칙 목록 조회",
        "operationId": "getAutoClassifyRules",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/X-User-Department" }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/AutoClassifyRuleResponse" } },
                "example": [
                  {
                    "id": 1,
                    "ruleName": "채용 관련 자동 분류",
                    "conditions": { "titleContains": "채용", "formName": "채용요청", "drafterDept": null, "drafterName": null },
                    "targetFolderId": 3,
                    "targetFolderName": "채용문서함",
                    "isActive": true,
                    "sortOrder": 1
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "tags": ["자동분류규칙"],
        "summary": "자동 분류 규칙 생성",
        "operationId": "createAutoClassifyRule",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/X-User-Department" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AutoClassifyRuleCreateRequest" },
              "example": {
                "ruleName": "채용 관련 자동 분류",
                "conditions": { "titleContains": "채용", "formName": "채용요청", "drafterDept": null, "drafterName": null },
                "targetFolderId": 3,
                "isActive": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성 성공",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AutoClassifyRuleResponse" } } }
          }
        }
      }
    },
    "/approval/auto-classify-rules/{id}": {
      "put": {
        "tags": ["자동분류규칙"],
        "summary": "자동 분류 규칙 수정",
        "operationId": "updateAutoClassifyRule",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/PathId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AutoClassifyRuleUpdateRequest" },
              "example": {
                "ruleName": "채용 관련 수정됨",
                "conditions": { "titleContains": "채용", "formName": null, "drafterDept": "인사팀", "drafterName": null },
                "targetFolderId": 3,
                "isActive": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 성공",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AutoClassifyRuleResponse" } } }
          }
        }
      },
      "delete": {
        "tags": ["자동분류규칙"],
        "summary": "자동 분류 규칙 삭제",
        "operationId": "deleteAutoClassifyRule",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/PathId" }
        ],
        "responses": { "204": { "description": "삭제 성공" } }
      }
    },
    "/approval/auto-classify-rules/{id}/toggle": {
      "patch": {
        "tags": ["자동분류규칙"],
        "summary": "자동 분류 규칙 활성/비활성 토글",
        "operationId": "toggleAutoClassifyRule",
        "parameters": [
          { "$ref": "#/components/parameters/X-User-Company" },
          { "$ref": "#/components/parameters/PathId" }
        ],
        "responses": { "200": { "description": "토글 성공" } }
      }
    },
    "/approval/auto-classify-rules/reorder": {
      "put": {
        "tags": ["자동분류규칙"],
        "summary": "자동 분류 규칙 순서 변경",
        "operationId": "reorderAutoClassifyRules",
        "parameters": [ { "$ref": "#/components/parameters/X-User-Company" } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReorderRequest" },
              "example": { "orderList": [ { "id": 2, "sortOrder": 1 }, { "id": 1, "sortOrder": 2 } ] }
            }
          }
        },
        "responses": { "200": { "description": "변경 성공" } }
      }
    },
...
  },
  "components": {
    "parameters": {
      "X-User-Company": {
        "name": "X-User-Company",
        "in": "header",
        "required": true,
        "schema": { "type": "string", "format": "uuid" },
        "description": "회사 ID (UUID)",
        "example": "550e8400-e29b-41d4-a716-446655440000"
      },
...
      "X-User-Department": {
        "name": "X-User-Department",
        "in": "header",
        "required": true,
        "schema": { "type": "integer", "format": "int64" },
        "description": "부서 ID",
        "example": 10
      },
...
      "PathId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64" },
        "description": "리소스 ID",
        "example": 1
      },
...
    },
    "schemas": {
...
      "ReorderItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64", "description": "리소스 ID" },
          "sortOrder": { "type": "integer", "description": "변경할 순서" }
        }
      },
      "ReorderRequest": {
        "type": "object",
        "properties": {
          "orderList": { "type": "array", "items": { "$ref": "#/components/schemas/ReorderItem" } }
        }
      },
...
      "AutoClassifyConditions": {
        "type": "object",
        "properties": {
          "titleContains": { "type": "string", "description": "제목 포함 키워드" },
          "formName": { "type": "string", "description": "양식명" },
          "drafterDept": { "type": "string", "description": "기안자 부서" },
          "drafterName": { "type": "string", "description": "기안자 이름" }
        }
      },
      "AutoClassifyRuleCreateRequest": {
        "type": "object",
        "properties": {
          "ruleName": { "type": "string", "description": "규칙명" },
          "conditions": { "$ref": "#/components/schemas/AutoClassifyConditions" },
          "targetFolderId": { "type": "integer", "format": "int64", "description": "대상 문서함 ID" },
          "isActive": { "type": "boolean" }
        }
      },
      "AutoClassifyRuleUpdateRequest": {
        "type": "object",
        "properties": {
          "ruleName": { "type": "string" },
          "conditions": { "$ref": "#/components/schemas/AutoClassifyConditions" },
          "targetFolderId": { "type": "integer", "format": "int64" },
          "isActive": { "type": "boolean" }
        }
      },
      "AutoClassifyRuleResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "ruleName": { "type": "string" },
          "conditions": { "$ref": "#/components/schemas/AutoClassifyConditions" },
          "targetFolderId": { "type": "integer", "format": "int64" },
          "targetFolderName": { "type": "string" },
          "isActive": { "type": "boolean" },
          "sortOrder": { "type": "integer" }
        }
      },
...
    }
  }
}